Add files via upload
Q1. Write all the conventions being followed while declaring a variable.
    Answer: 
     i. Variable names should be lowercase, and words should be separated by underscores.
    ii. Blank space cannot be used as variable name.
   iii. Variable names are case sensitive.  
    iv. Special character cannot be used in variable name.

Q2. What will happen if we declare a restricted keyword as a variable?
    Answer:
    If we declare a restricted keyword as a variable, then we will face syntax Error.

Q3. Can we actually declare a string as a variable name?
     Answer:
    Yes, we can declare a string as a variable name.

Q4. Is it possible for us to declare "_" as a variable? if so, then write an example of it.
     Answer: 
       Yes, it is possible.
     
     _="Prem"
      print(_)

Q5. Using an example, explain how the variables in python are dynamic in nature.
      Answer:
      a=9
      a=2
      print(a)
     
    –> Initally, variable a was assign with value 9 but after again a is assign with value 
      2  which make sure variable is dynamic.